# Conflicts:
#	Barotrauma/BarotraumaShared/BarotraumaShared.projitems
#	Barotrauma/BarotraumaShared/Source/Characters/AI/Objectives/AIObjectiveCombat.cs

+++ Barotrauma/BarotraumaShared/Source/Characters/CharacterNetworking.cs
@@ -358,7 +358,7 @@ namespace Barotrauma
                             }
                             break;
                         case 3:
-                            LimbType grabLimb = (LimbType)msg.ReadUInt16();
+                            LimbType grabLimb = (LimbType)msg.ReadByte();
                             if (c.Character != this)
                             {
 #if DEBUG
@@ -449,7 +449,7 @@ namespace Barotrauma
                     if (AnimController is HumanoidAnimController)
                     {
                         tempBuffer.Write(((HumanoidAnimController)AnimController).Crouching);
-                        tempBuffer.Write((UInt16)AnimController.GrabLimb);
+                        tempBuffer.Write((byte)AnimController.GrabLimb);
                     }
 
                     bool hasAttackLimb = AnimController.Limbs.Any(l => l != null && l.attack != null);
